+++ shared/src/androidMain/kotlin/com/electra/template/data/storage/SettingsFactory.android.kt
@@ -1,9 +1,0 @@
-package com.electra.template.data.storage
-
-import android.content.Context
-import com.russhwolf.settings.Settings
-import com.russhwolf.settings.SharedPreferencesSettings
-
-actual class SettingsFactory(private val context: Context) {
-    actual fun create(name: String): Settings = SharedPreferencesSettings(context.getSharedPreferences(name, Context.MODE_PRIVATE))
-}