+++ shared/src/commonMain/kotlin/com/electra/template/core/di/PlatformModule.kt
@@ -1,5 +1,0 @@
-package com.electra.template.core.di
-
-import org.koin.core.module.Module
-
-expect val platformModule: Module