+++ shared/src/commonMain/kotlin/com/electra/template/data/storage/SettingsFactory.kt
@@ -2,6 +2,6 @@ package com.electra.template.data.storage
 
 import com.russhwolf.settings.Settings
 
-expect class SettingsFactory {
+interface SettingsFactory {
     fun create(name: String = "app"): Settings
 }