+++ shared/src/iosMain/kotlin/com/electra/template/data/storage/IosSettingsFactory.kt
@@ -4,7 +4,7 @@ import com.russhwolf.settings.NSUserDefaultsSettings
 import com.russhwolf.settings.Settings
 import platform.Foundation.NSUserDefaults
 
-actual class SettingsFactory {
-    actual fun create(name: String): Settings =
+class IosSettingsFactory : SettingsFactory {
+    override fun create(name: String): Settings =
         NSUserDefaultsSettings(NSUserDefaults(suiteName = name) ?: NSUserDefaults.standardUserDefaults)
 }