+++ shared/src/iosMain/kotlin/com/electra/template/ios/KoinInitializer.kt
@@ -1,14 +1,14 @@
 package com.electra.template.ios
 
 import com.electra.template.core.di.coreModule
-import com.electra.template.core.di.platformModule
+import com.electra.template.core.di.iosPlatformModule
 import com.electra.template.data.todo.todoModule
 import org.koin.core.context.startKoin
 
 class KoinInitializer {
     fun doInit() {
         startKoin {
-            modules(coreModule, platformModule, todoModule)
+            modules(coreModule, iosPlatformModule, todoModule)
         }
     }
 }